Development session 08/03/2023

+++ ecm1410_coursework/src/socialmedia/Endorsement.java
@@ -1,7 +1,7 @@
 package socialmedia;
 public class Endorsement extends Post{
-    private long parentID;
-    public Endorsement(String handle, long PostID){
+    private int parentID;
+    public Endorsement(String handle, int PostID){
         parentID = PostID;
         this.PostID = numberofPosts++;
         this.handle = handle;
